Move CallLog component to misc directory
